sailor/requirements.txt
aiohttp==3.9.4
#anydata_sdk==0.0.3.791077
APScheduler==3.10.4
async_timeout==4.0.3
coverage==7.3.2
fastapi==0.110.3
jieba==0.42.1
jinja2==3.1.2
kafka-python==2.0.2
langchain==0.1.0
langchain-community==0.0.16
langchain-openai==0.0.5
numpy==1.26.3
opensearch-py==2.4.2
pandas==2.1.2
pyahocorasick==2.0.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.1.0
pylint==3.0.2
redis==5.0.4
requests==2.31.0
sqlparse==0.5.0
sql_metadata==2.12.0
scikit-learn==1.3.2
sse-starlette==1.8.2
starlette==0.37.2
tabulate==0.9.0
urllib3==2.1.0
uvicorn==0.23.2
#websocket==0.2.1
#websockets==12.0
#websocket-client==1.7.0
xmlrunner==1.7.7






